requirements/apps.txt
django==1.3.1
South==0.7.5
gunicorn==0.12.2
python-dateutil==1.5
django-pagination==1.0.7
lxml==2.3
twilio==2.0.6
tropo-webapi-python==0.1.1
django-sorting==0.1
psycopg2==2.4.1
django-rosetta==0.6.0
django-celery==2.2.4
celery==2.2.7
python-twitter==0.8.2
feedparser==5.0.1
gevent==0.13.6
greenlet==0.3.1
distribute>=0.6.19
couchdbkit==0.5.4
rapidsms-xforms==0.4.1
django-uni-form==0.8.0
python-digest==1.7
django-digest==1.13
django-timezones==0.2
pytz==2012c
openpyxl==1.5.8
unidecode==0.04.9
requests==0.13.1
certifi==0.0.8
oauthlib==0.1.3
chardet==1.0.1
rsa==3.1.1
pyasn1==0.1.3
django-selectable==0.2.0
rapidsms-threadless-router==0.1.3
rapidsms-tropo==0.1.1
-e git+https://github.com/dimagi/rapidsms.git#egg=rapidsms
-e git+https://github.com/adammck/djtables#egg=djtables
-e git+https://github.com/dimagi/couchlog.git#egg=couchlog
-e git+https://github.com/dimagi/auditcare.git#egg=auditcare
-e git+https://github.com/dimagi/dimagi-utils.git#egg=dimagi_utils
-e git+https://github.com/caktus/rapidsms-decisiontree-app.git@develop#egg=decisiontree
-e git+http://github.com/mvpdev/django-eav.git#egg=django-eav-0.9.2
